ml/requirements.txt
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
pyarrow>=14.0.0
psycopg2-binary>=2.9.0
matplotlib>=3.7.0
gradio>=4.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
